Moving the SkillGenerator to the personnel.generator folder
